+++ api/.flaskenv
@@ -1,0 +1,3 @@
+FLASK_APP=api.py
+FLASK_ENV=development
+SCRAPY_SETTINGS_MODULE=spider.spider.settings
